Update attribute configuration ouytside of value converter (#199)
